saved work pose graph node
